Update principal_engineer.md
Seems like an example of how we confuse lead in lead engineer vs lead in product engineering lead.

Correct me if I'm wrong but this line is more about consulting with the team engineering leads (PEL or just EL in my case) and the PMs on the team.  Not staff engineers specifically.